Improve code writing
Lots of modifications were made. The first one is to replace common
expressions with lombok type inference feature "val", the other one is
to replace written get/set functions with the @data annotation from
lombok as well.

Other modifications were made to simplify the code and to be ease to
modify in the future, adding more expressions on the metrics collector
if needed.

Signed-off-by: Rafael Campos Nunes <rafaelnunes@engineer.com>